+++ mvu_layer/example/pubspec.yaml
@@ -51,9 +51,6 @@ dev_dependencies:
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
 
-dependency_overrides:
-  analyzer: 5.12.0
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
